Settings for Gazetter
