+++ web-ui/src/main/resources/catalog/locales/en-admin.json
@@ -1174,6 +1174,9 @@
     "allLanguage-help": "Display all languages",
     "oneLanguage": "One",
     "oneLanguage-help": "Choose one language to edit using the selector",
+    "ui-mod-geocoder": "Gazetteer",
+    "ui-behaviour": "Behaviour (default)",
+    "ui-geocoder-help": "Insert the url of Gazetteer service and the behaviour (alt or default)",
     "ui-mapExtent": "DEPRECATED! Map Extent",
     "ui-mapExtent-help": "DEPRECATED! Extent coordinates to use for maps other that viewer and search maps.",
     "ui-mapBackgroundLayer": "DEPRECATED! Map Background Layer",
